Django_API/Django_API/requirements.txt
asgiref==3.2.10
boto==2.49.0
boto3==1.14.43
botocore==1.17.43
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-crispy-forms==1.9.2
djangorestframework==3.11.1
docutils==0.15.2
fasttext==0.9.2
gunicorn==19.0.0
idna==2.10
jmespath==0.10.0
joblib==0.16.0
nltk==3.2
numpy==1.19.1
pandas==1.1.0
psycopg2-binary==2.8.5
pybind11==2.5.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
smart-open==2.1.0
sqlparse==0.3.1
urllib3==1.25.10
